Update SCR service and docs
